Add phase options. Change name to label. Add ord
Standardised all field names to "label" in order to simplify things.

Added a new column to many tables called "ord". The intention of this
column is to use it to sort the order of the rows in a way that remains
consistent as new data is added and existing data is modified. This is
requried because all IDs have been changed to UUIDs which can't be
sorted as we did previously with integer IDs.